Follow the systemd ideology.
Introduce two new concepts: run-files and "inner layer".
First is a file containing some argument set, which can be written by
anybody (i. e. this script itself) and read by the script when it is
passed a hidden "runfile" target.
Then, make the script - when called by the user - not start the daemons
directly; instead, write computed parameters to the run-file and invoke
systemd to start the unit. Systemd unit, in turn, starts the same script
but in a hidden "inner" mode and with "runfile" target. Inner mode makes
the script to directly start daemons under systemd's cgroup and etc.

Directly starting systemd unit is still supported as "runfile" targets
defaults to "bootup" if no run-file is present.

Sleeping shall be supported now properly as the
net-auto-wireless-suspend.service unit which invokes
netcfg-auto-wireless is no longer an owner of resumed netcfg processes.

This works with anything but changing mode by reconfiguring.